requirements.txt
easyocr>=1.7.1
ffmpeg-python>=0.2.0
paddleocr>=2.7.0
paddlepaddle>=2.5.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
opencc-python-reimplemented>=0.1.7 